chore: move logger code
